Move refreshIndexPattern to service
